requirements.txt
fastapi==0.115.12
uvicorn==0.33.0
requests==2.32.3
openai==1.76.0
python-dotenv==1.0.1